+++ R/cluster_stats.R
@@ -30,38 +30,31 @@ setMethod(
     df <- data.frame(size=clust_size(object), 
                      row.names = names(clust_size(object)))
     
+    print_msg('Splitting expression data', msg_type="DEBUG")
     gene_clust <- as.factor(gene_cluster(object))
+    df_split <- split(as.data.frame(object@data), gene_clust)
+    
+    print_msg('Computing sum', msg_type="DEBUG")
+    tmp <- unlist(lapply(df_split, sum)) 
+    tmp_2 <-  clust_size(object)
+
+
+    df$sum_by_row <- tmp[row.names(df)] / tmp_2[row.names(df)]
+    
+    print_msg('Computing sd.', msg_type="DEBUG")
     df_split <- split(object@data, gene_clust)
     
-    tmp <- unlist(lapply(df_split, sum))
-    df$sum_count <- tmp[row.names(df)]
+    print_msg('Computing var.', msg_type="DEBUG")
+    tmp <- unlist(lapply(df_split, stats::var)) 
+    df$var <- tmp[row.names(df)]
     
-    tmp <- unlist(lapply(df_split, stats::var))
-    df$var_total <- tmp[row.names(df)]
-    
+
     tmp <- unlist(lapply(df_split, stats::sd))
-    df$sd_total <- tmp[row.names(df)]
-    
-    all_dot_prod <- vector()
-    
-    for (i in 1:length(object@gene_clusters)) {
-      
-      print_msg(paste0("Computing dot product for cluster: ", i), 
-                msg_type = "DEBUG")
-      cur_clust <- object@data[object@gene_clusters[[i]],]
-      cur_clust[cur_clust >= 1] <- 1
-      cur_clust[cur_clust < 1] <- 0
-      cur_dot_prod <- cur_clust %*% t(cur_clust)
-      diag(cur_dot_prod) <- NA
-      cur_dot_prod_median_of_max <-median(apply(cur_dot_prod, 
-                                                1, 
-                                                max, 
-                                                na.rm = T))
-      all_dot_prod[i] <- cur_dot_prod_median_of_max
-      
-    }
-    
-    df$dot_prod <- all_dot_prod
+    df$sd <- tmp[row.names(df)]
+ 
+    print_msg('Computing cv.', msg_type="DEBUG")
+    tmp <- unlist(lapply(df_split, stats::sd)) / unlist(lapply(df_split, mean))
+    df$cv <- tmp[row.names(df)]
     
     return(df)